refactor: move CodePoint to Core namespace
