--- src/Danaus/Url/Encoding.cs
@@ -1,6 +1,0 @@
-namespace Danaus.Url;
-
-public enum Encoding
-{
-    UTF8
-}
